+++ tests/test_fragment_select.py
@@ -5,6 +5,7 @@ from bs4 import BeautifulSoup
 
 from wikipedia_ql.fragment import Fragment
 from wikipedia_ql.selectors import text, text_group, sentence, section, css, alt, page
+from wikipedia_ql.parser import Parser
 
 def make_fragment(html):
     # Because fragment is Wikipedia-oriented and always looks for this div :shrug:
@@ -13,28 +14,27 @@ def make_fragment(html):
 def h(html):
     return re.sub(r'\s+<', '<', re.sub(r'>\s+', '>', html))
 
+def sel(text):
+    return Parser().parse_selector(text)
+
 def apply(fragment, selector, simplify=False):
+    selector = sel(selector)
+    fragment = make_fragment(fragment)
+    print(selector)
     if simplify:
-        return [h(str(f.soup)) for f in fragment._select(selector)]
+        return [h(str(f.soup)) for f in fragment.select(selector)]
     else:
-        return [str(f.soup) for f in fragment._select(selector)]
+        return [str(f.soup) for f in fragment.select(selector)]
 
 def test_select_text():
-    def select(html, selector):
-        return apply(make_fragment(html), text(pattern=selector))
-
-    assert select("""
+    assert apply("""
             <h2>Header</h2>
             <p>Paragraph with a <a href="foo">link</a>. And <b>bold</b> text.</p>
             """,
-            'with a link') == ['<span>with a <a href="foo">link</a></span>']
-
+            'text:matches("with a link")') == ['<span>with a <a href="foo">link</a></span>']
 
 def test_select_section():
-    def select(fragment, **args):
-        return apply(fragment, section(**args), simplify=True)
-
-    fragment = make_fragment("""
+    fragment = """
         <section>
             <h2>Section1</h2>
             <p>Text1</p>
@@ -51,9 +51,9 @@ def test_select_section():
             <h2>Section2</h2>
             <p>Text5</p>
         </section>
-        """)
+        """
 
-    assert select(fragment, heading='Section1') == [
+    assert apply(fragment, 'section[heading="Section1"]', simplify=True) == [
         h("""
         <div>
             <h2>Section1</h2>
@@ -80,10 +80,9 @@ def test_select_section():
             <p>Text4</p>
         </div>
         """)
-
     ]
 
-    assert select(fragment, heading='Section1.1') == [
+    assert apply(fragment, 'section[heading="Section1.1"]', simplify=True) == [
         h("""
         <div>
                 <h3>Section1.1</h3>
@@ -92,13 +91,30 @@ def test_select_section():
         """)
     ]
 
-    assert select(fragment, heading='Section2') == [
+    assert apply(fragment, 'section[heading="Section2"]', simplify=True) == [
         h("""
         <div>
         <h2>Section2</h2>
         <p>Text5</p>
         </div>
         """)
+    ]
+
+    assert apply(fragment, 'section:first', simplify=True) == [
+        h("""
+        <div>
+            <h2>Section1</h2>
+            <p>Text1</p>
+            <section>
+                <h3>Section1.1</h3>
+                <p>Text2</p>
+            </section>
+            <section>
+                <h3>Section1.2</h3>
+                <p>Text4</p>
+            </section>
+        </div>
+        """),
     ]
     # TODO:
     # all
@@ -107,50 +123,41 @@ def test_select_section():
     # text pattern (include, start/stop with, regexp)
 
 def test_select_css():
-    def select(fragment, css_selector):
-        return apply(fragment, css(css_selector=css_selector), simplify=True)
-
-    fragment = make_fragment("""
+    fragment = """
         <h2>Section1</h2>
         <p>Text1</p>
         <ul>
             <li><a class="first">Link1</a></li>
             <li class="item"><a class="second">Link2</a>text</li>
         </ul>
-        """)
+        """
 
-    assert select(fragment, '.item') == [
+    assert apply(fragment, '.item') == [
         h("""
         <li class="item"><a class="second">Link2</a>text</li>
         """)
     ]
 
 def test_select_sentence():
-    def select(fragment, pattern):
-        return apply(fragment, sentence(pattern=pattern))
-
-    fragment = make_fragment(
-        """
+    fragment = """
         <p>This is <b>sentence</b> one. This is phrase <a href="foo">two.</a> This is NOT</p>
         <p>sentence three, probably!</p>
         """
-    )
 
-    assert select(fragment, 'one') == ['<span>This is <b>sentence</b> one.</span>']
-    assert select(fragment, 'This.+two') == ['<span>This is phrase <a href="foo">two.</a></span>']
-    assert select(fragment, 'This.+three') == []
-    assert select(fragment, None) == [
+    assert apply(fragment, 'sentence:contains("one")') == ['<span>This is <b>sentence</b> one.</span>']
+    assert apply(fragment, 'sentence:contains("This.+two")') == ['<span>This is phrase <a href="foo">two.</a></span>']
+    assert apply(fragment, 'sentence:contains("This.+three")') == []
+    assert apply(fragment, 'sentence') == [
         '<span>This is <b>sentence</b> one.</span>',
         '<span>This is phrase <a href="foo">two.</a></span>',
         'This is NOT',              # the whole text after </a>
         'sentence three, probably!' # takes the whole paragraph, that's why no span
     ]
+    assert apply(fragment, 'sentence:first') == ['<span>This is <b>sentence</b> one.</span>']
+
 
 def test_select_nested():
-    def select(fragment, selector):
-        return [str(f.soup) for f in fragment.select(selector)]
-
-    fragment = make_fragment("""
+    fragment = """
         <section>
             <h2>Section1</h2>
             <p>Text1</p>
@@ -159,58 +166,49 @@ def test_select_nested():
                 <li class="item"><a class="second">Link2</a>text</li>
             </ul>
         </section>
-        """)
+        """
 
-    selector = section(heading='Section1', nested=css(css_selector='ul', nested=text(pattern='Li...')))
-    assert select(fragment, selector) \
+    selector = 'section[heading="Section1"] >> ul >> text:matches("Li...")'
+    assert apply(fragment, selector) \
         == ['<a class="first">Link1</a>', '<a class="second">Link2</a>']
 
     # edge case: css selector of top-level node:
-    selector = text(pattern='Link1', nested=css(css_selector='a'))
-    assert select(fragment, selector) == ['<a class="first">Link1</a>']
+    assert apply(fragment, 'text:matches("Link1") >> a') == ['<a class="first">Link1</a>']
 
 def test_select_alt():
-    def select(fragment, *selectors):
-        return apply(fragment, alt(*selectors), simplify=True)
-
-    fragment = make_fragment("""
+    fragment = """
         <h2>Section1</h2>
         <p>Text1</p>
         <ul>
             <li><a class="first">Link1</a></li>
             <li class="item"><a class="second">Link2</a>text</li>
         </ul>
-        """)
+        """
 
-    assert select(fragment, text(pattern='Link1'), css(css_selector='a.second')) == \
+    assert apply(fragment, '{ text:matches("Link1"); a.second }') == \
         ['<a class="first">Link1</a>', '<a class="second">Link2</a>']
 
 def test_select_text_group():
-    def select(fragment, group):
-        return apply(fragment, text_group(group=group))
+    fragment = '<p>Some paragraph with <b>empasis</b> and <a href="#foo">link</a></p>'
 
-    fragment = make_fragment(
-        '<p>Some paragraph with <b>empasis</b> and <a href="#foo">link</a></p>'
-    )
-
-    assert select(fragment.select(text(pattern=r'with (\S+ and) li(.*)')), 1) == [
+    assert apply(fragment, r'text:matches("with (\S+ and) li(.*)") >> text-group[group=1]') == [
         '<span><b>empasis</b> and</span>'
     ]
 
     # When not after text
     with pytest.raises(ValueError, match='text-group is only allowed after text'):
-        select(fragment, 1)
+        apply(fragment, 'text-group[group=1]')
 
     # When slice index is out of range
-    assert select(fragment.select(text(pattern=r'with (\S+ and) li(.*)')), 10) == []
+    assert apply(fragment, r'text:matches("with (\S+ and) li(.*)") >> text-group[group=10]') == []
 
     # Named groups
-    assert select(fragment.select(text(pattern=r'with (?P<group1>\S+ and) li(.*)')), 'group1') == [
+    assert apply(fragment, r'text:matches("with (?P<group1>\S+ and) li(.*)") >> text-group[group=group1]') == [
         '<span><b>empasis</b> and</span>'
     ]
 
     # Non-existent group:
-    assert select(fragment.select(text(pattern=r'with (?P<group1>\S+ and) li(.*)')), 'group2') == []
+    assert apply(fragment, r'text:matches("with (?P<group1>\S+ and) li(.*)") >> text-group[group=group2]') == []
 
 def test_select_page():
     fragment = make_fragment("""
@@ -222,8 +220,8 @@ def test_select_page():
         </ul>
         """)
 
-    assert str(fragment.select(page()).items[0].soup) == str(fragment.soup)
+    assert str(fragment.select(sel('page')).items[0].soup) == str(fragment.soup)
 
-    assert str(fragment.select(css(css_selector='li', nested=page())).items[0].soup) == str(fragment.soup)
+    assert str(fragment.select(sel('li >> page')).items[0].soup) == str(fragment.soup)
 
-# TODO: (laterz!) wikitable, infobox, ...
+# TODO: (laterz!) infobox, ...